quiz-website/app/requirements.txt
flask
pillow
pdf2image
mysql-connector-python
gunicorn